refactor(desktop): split EditProjectPane.tsx by feature seam
Extracts lifecycle helpers, error mapping, and the channel-drift and
sources sub-panels (each with its own hook) into colocated modules,
matching the three concerns already named in the file's own doc
comment (US3/US4/WP-008-C). Public import path and props are
unchanged; no behaviour change.

Refs #1000
